Update snmptt to add file locking support
File locking was added to spool files in snmptt v1.5 when processed by snmptthandler or snmpttthandler-embedded.  This locking using flock or fcntl was not added to the main snmptt program. 

File locking on Linux is advisory in nature rather than explicit in Windows.  On Linux hosts with file locking support, the subsequent read operation by snmptt on a spool file was not obeying the file lock and appears to have been allowing snmptt to read spool files that were not totally written when a large number of traps occurred.   This occurred because snmptt itself did not check for the file lock before ingesting the spool file. 

This update ports the code from snmptthandler-embedded into snmptt and attempts to acquire a shared lock on each spool file.  Should it be unable to acquire the lock (because the handler is still writing it), it will loop every 200ms waiting for the lock to become available until it reaches a set timeout (default 5s).  Should timeout be reached, the attempt is logged and the file is read as-is.